changed Activity names, moved them into new folder.
