+++ app/src/main/AndroidManifest.xml
@@ -18,7 +18,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".Home">
+        <activity android:name=".Activities.HomeActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -34,26 +34,26 @@
             android:value="@integer/google_play_services_version" />
 
         <activity
-            android:name=".AR_Page"
+            android:name=".Activities.AR_Page_Activity"
             android:label="@string/app_name" />
-        <activity android:name=".Building_List" >
+        <activity android:name=".Activities.BuildingListActivity" >
             <intent-filter>
                 <action android:name="android.intent.action.PICK_ACTIVITY." />
 
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
         </activity>
-        <activity android:name=".Events" />
-        <activity android:name=".Contact" />
+        <activity android:name=".Activities.EventsActivity" />
+        <activity android:name=".Activities.ContactActivity" />
         <activity android:name=".Maps_Activity"/>
         <activity
-            android:name=".Create_events"
+            android:name=".Activities.CreateEventsActivity"
             android:label="@string/title_activity_create_events"
             android:theme="@style/AppTheme.NoActionBar" />
         <activity
             android:name=".map_main"
             android:label="@string/title_activity_map_main"
-            android:theme="@style/AppTheme"></activity>
+            android:theme="@style/AppTheme"/>
     </application>
 
 </manifest>
