+++ app/src/main/java/com/example/leolam/myapplication/Activities/AR_Page_Activity.java
@@ -1,9 +1,11 @@
-package com.example.leolam.myapplication;
+package com.example.leolam.myapplication.Activities;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-public class AR_Page extends AppCompatActivity {
+import com.example.leolam.myapplication.R;
+
+public class AR_Page_Activity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
