+++ app/src/main/java/com/example/leolam/myapplication/Activities/ContactActivity.java
@@ -1,4 +1,4 @@
-package com.example.leolam.myapplication;
+package com.example.leolam.myapplication.Activities;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -8,7 +8,9 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
 
-public class Contact extends AppCompatActivity {
+import com.example.leolam.myapplication.R;
+
+public class ContactActivity extends AppCompatActivity {
     private EditText mEditTextSubject;
     private EditText mEditTextMessage;
