+++ app/src/main/java/com/example/leolam/myapplication/Activities/EventsActivity.java
@@ -1,22 +1,14 @@
-package com.example.leolam.myapplication;
+package com.example.leolam.myapplication.Activities;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.Manifest;
-import android.annotation.TargetApi;
 import android.content.Intent;
-import android.content.pm.PackageManager;
-import android.icu.util.Freezable;
-import android.os.Build;
-import android.provider.MediaStore;
-import android.support.v4.app.ActivityCompat;
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
-import android.widget.TextView;
 
-public class Events extends AppCompatActivity {
+import com.example.leolam.myapplication.R;
+
+public class EventsActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -24,11 +16,11 @@ public class Events extends AppCompatActivity {
         setContentView(R.layout.activity_events);
         Button Create_events = (Button)findViewById(R.id.create_event);
 
-        //Contact Button to Contact Page
+        //ContactActivity Button to ContactActivity Page
         Create_events.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent signup = new Intent(Events.this, Create_events.class);
+                Intent signup = new Intent(EventsActivity.this, CreateEventsActivity.class);
                 startActivity(signup);
             }
         });
