+++ app/src/main/res/layout/activity_building__list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#ccd9c8"
-    tools:context=".Building_List">
+    tools:context=".Activities.BuildingListActivity">
 
 
     <Spinner
