+++ app/src/main/res/layout/activity_contact.xml
@@ -7,7 +7,7 @@
     android:orientation="vertical"
     android:padding="16dp"
     android:background="#d4fcc4"
-    tools:context=".Contact">
+    tools:context=".Activities.ContactActivity">
 
 
     <TextView
