+++ app/src/main/res/layout/activity_create_events.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#ccd9c8"
-    tools:context=".Create_events">
+    tools:context=".Activities.CreateEventsActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
