+++ app/src/main/res/layout/activity_events.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:background="#ccd9c8"
     android:layout_height="match_parent"
-    tools:context=".Events">
+    tools:context=".Activities.EventsActivity">
 
     <Button
         android:id="@+id/create_event"
