+++ app/src/main/res/layout/activity_home.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:background="@drawable/background"
     android:layout_height="match_parent"
-    tools:context=".Home">
+    tools:context=".Activities.HomeActivity">
 
     <ImageView
         android:id="@+id/imageView"
