+++ app/src/main/res/layout/content_create_events.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".Create_events"
+    tools:context=".Activities.CreateEventsActivity"
     tools:showIn="@layout/activity_create_events">
 
     <EditText
